requirements.txt
Django>=5.0.0
django-allauth>=0.50.0
pillow>=11.0.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1
python-dotenv>=1.0.0
stripe>=5.0.0
django-countries>=7.0